+++ pkg/webhook/connectorcatalog_webhook.go
@@ -20,6 +20,7 @@ package webhook
 import (
 	"context"
 	"fmt"
+
 	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
